+++ library/src/commonMain/kotlin/org/jraf/klibnotion/internal/model/property/value/formula/StringFormulaPropertyValueImpl.kt
@@ -29,5 +29,5 @@ import org.jraf.klibnotion.model.property.value.StringFormulaPropertyValue
 internal data class StringFormulaPropertyValueImpl(
     override val id: String,
     override val name: String,
-    override val value: String,
+    override val value: String?,
 ) : StringFormulaPropertyValue