+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add support for [retrieve block API](https://developers.notion.com/reference/retrieve-a-block).
 - Add support for [update block API](https://developers.notion.com/reference/update-a-block).
+- Add support for [update database API](https://developers.notion.com/reference/update-a-database).
 
 ## v1.6.0 (2021-07-25)