+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next version (????-??-??)
 
+## v1.7.0 (2021-09-10)
+
 - Add support for [retrieve block API](https://developers.notion.com/reference/retrieve-a-block).
 - Add support for [update block API](https://developers.notion.com/reference/update-a-block).
 - Add support for [update database API](https://developers.notion.com/reference/update-a-database).